refactor: move custom sources to providers
