+++ source/chapter.go
@@ -1,12 +1,10 @@
 package source
 
 import (
-	"errors"
 	"fmt"
 	"github.com/metafates/mangal/config"
 	"github.com/metafates/mangal/util"
 	"github.com/spf13/viper"
-	lua "github.com/yuin/gopher-lua"
 	"strings"
 	"sync"
 )
@@ -23,30 +21,6 @@ type Chapter struct {
 
 func (c *Chapter) String() string {
 	return c.Name
-}
-
-func chapterFromTable(table *lua.LTable, manga *Manga, index uint16) (*Chapter, error) {
-	name := table.RawGetString("name")
-
-	if name.Type() != lua.LTString {
-		return nil, errors.New("type of field \"name\" should be string")
-	}
-
-	url := table.RawGetString("url")
-	if url.Type() != lua.LTString {
-		return nil, errors.New("type of field \"url\" should be string")
-	}
-
-	chapter := &Chapter{
-		Name:  strings.TrimSpace(name.String()),
-		URL:   strings.TrimSpace(url.String()),
-		Manga: manga,
-		Index: index,
-		Pages: []*Page{},
-	}
-
-	manga.Chapters = append(manga.Chapters, chapter)
-	return chapter, nil
 }
 
 // DownloadPages downloads the Pages contents of the Chapter.
