+++ source/page.go
@@ -3,11 +3,8 @@ package source
 import (
 	"errors"
 	"github.com/metafates/mangal/constant"
-	lua "github.com/yuin/gopher-lua"
 	"io"
 	"net/http"
-	"strconv"
-	"strings"
 )
 
 type Page struct {
@@ -17,36 +14,6 @@ type Page struct {
 	SourceID  string `json:"source_id"`
 	Contents  io.ReadCloser
 	Chapter   *Chapter
-}
-
-func pageFromTable(table *lua.LTable, chapter *Chapter) (*Page, error) {
-	url := table.RawGetString("url")
-
-	if url.Type() != lua.LTString {
-		return nil, errors.New("type of field \"url\" should be string")
-	}
-
-	index := table.RawGetString("index")
-
-	if index.Type() != lua.LTNumber {
-		return nil, errors.New("type of field \"index\" should be number")
-	}
-
-	num, err := strconv.ParseUint(index.String(), 10, 16)
-
-	if err != nil {
-		return nil, errors.New("index must be an unsigned 16 bit integer")
-	}
-
-	page := &Page{
-		URL:       strings.TrimSpace(url.String()),
-		Index:     uint16(num),
-		Chapter:   chapter,
-		Extension: ".jpg",
-	}
-
-	chapter.Pages = append(chapter.Pages, page)
-	return page, nil
 }
 
 func (p *Page) Download() error {
